config/module: use go-getter
